{"schema":"decodex/commit/1","summary":"split automation support modules","authority":"manual"}
