+++ automations/radar/scripts/github/contract_support/__init__.py
@@ -1,0 +1,1 @@
+"""Modular contract helpers for GitHub automation scripts."""
